+++ sqlparser/lib/utils/find_referenced_tables.dart
@@ -162,7 +162,7 @@ extension FindReferenceAnalysis on SqlEngine {
         schemaSupport: schemaReader);
 
     AstPreparingVisitor(context: context).start(clone);
-    clone.acceptWithoutArg(ColumnResolver(context));
+    clone.accept(ColumnResolver(context), const ColumnResolverContext());
 
     return scope.addedTables;
   }